% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtros.R
\name{agrupamento_dados}
\alias{agrupamento_dados}
\title{Realizar agrupamento e soma de dados}
\usage{
agrupamento_dados(dados, campos_agrupar, campos_somar)
}
\arguments{
\item{dados}{Um data frame contendo os dados a serem agrupados e somados.}

\item{campos_agrupar}{Um vetor de caracteres contendo os nomes dos campos pelos quais os dados serão agrupados.}

\item{campos_somar}{Um vetor de caracteres contendo os nomes dos campos a serem somados.}
}
\value{
Um novo data frame com os dados agrupados e somados.
}
\description{
Esta função realiza o agrupamento e soma de dados com base nos campos especificados.
Ela utiliza as funções `group_by_at` e `summarise` do pacote dplyr para realizar o agrupamento e a soma dos campos especificados.
}
