% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suporte_server.R
\name{atualiza_filtro_data}
\alias{atualiza_filtro_data}
\title{Atualizar Filtro de Data}
\usage{
atualiza_filtro_data(database, fato_tipo)
}
\arguments{
\item{database}{Um objeto que contem uma conexão com o banco de dados.}

\item{fato_tipo}{O nome da tabela do banco de dados que contém as informações de data.}
}
\value{
Nenhum valor é retornado. A função apenas atualiza o filtro de data no aplicativo Shiny.
}
\description{
Usando a conexão com o banco armazenada no parâmetro `database`, e a tabela do parâmetro `fato_tipo`.
Esta função atualiza o filtro de data com base nos valores máximos e mínimos disponíveis nesta tabela do banco de dados.
}
