% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suporte_server.R
\name{atualiza_filtro_municipio}
\alias{atualiza_filtro_municipio}
\title{Atualizar Filtro de Município}
\usage{
atualiza_filtro_municipio(database, session, input)
}
\arguments{
\item{database}{Um objeto de conexão com o banco de dados.}

\item{session}{O objeto da sessão Shiny.}

\item{input}{O objeto de entrada Shiny.}
}
\value{
Nenhum valor é retornado. A função atualiza o filtro de seleção de municípios no aplicativo Shiny.
}
\description{
Usando a conexão com o banco armazenada no parâmetro `database` esta função atualiza o filtro de seleção de municípios com base nos inputs espaciais do aplicativo Shiny.
}
