% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suporte_server.R
\name{atualiza_filtro_n1_secao}
\alias{atualiza_filtro_n1_secao}
\title{Atualizar Filtro de Nível 1 - Seção}
\usage{
atualiza_filtro_n1_secao(database, session, input, fato_tipo)
}
\arguments{
\item{database}{Um objeto de conexão com o banco de dados.}

\item{session}{O objeto da sessão Shiny.}

\item{input}{O objeto de entrada Shiny.}

\item{fato_tipo}{O nome da tabela do banco de dados que contém as informações do fato.}
}
\value{
Nenhum valor é retornado. A função atualiza o filtro de seleção do Nível 1 (Seção) no aplicativo Shiny.
}
\description{
Usando a conexão com o banco armazenada no parâmetro `database`, e a tabela do parâmetro `fato_tipo`.
Esta função atualiza o filtro de seleção do Nível 1 (Seção) com base nos valores disponíveis nesta tabela do banco de dados.
}
