% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suporte_server.R
\name{atualiza_titulo_pagina}
\alias{atualiza_titulo_pagina}
\title{Atualizar Título da Página}
\usage{
atualiza_titulo_pagina(painel = NULL)
}
\arguments{
\item{painel}{O nome do painel selecionado.}
}
\value{
Nenhum valor é retornado. A função atualiza o título da página no aplicativo Shiny.
}
\description{
Esta função atualiza o título da página com base no painel selecionado no aplicativo Shiny.
}
