% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beautiful_table.R
\name{beautiful_table}
\alias{beautiful_table}
\title{Criar uma Tabela Bonita}
\usage{
beautiful_table(
  dados_brutos,
  campos_agrupar,
  campos_somar,
  campos_informacao = NULL,
  output_order = NULL
)
}
\arguments{
\item{dados_brutos}{Os dados brutos a serem resumidos e agrupados.}

\item{campos_agrupar}{Um vetor de caracteres especificando as colunas pelas quais os dados serão agrupados.}

\item{campos_somar}{Um vetor de caracteres especificando as colunas que serão somadas.}

\item{campos_informacao}{Um vetor de caracteres opcional especificando colunas adicionais a serem incluídas na tabela.}

\item{output_order}{Um vetor de caracteres opcional especificando a ordem das colunas na tabela de saída.
Se não for fornecido, a ordem será determinada pela concatenação de `campos_somar` e `campos_informacao`.}
}
\value{
Um tibble contendo os dados resumidos e agrupados em um formato de tabela bonita.
}
\description{
Esta função gera uma tabela bonita resumindo e agrupando dados com base em colunas especificadas.
A beleza da tabela está nos campos agrupados serem dispostos numa única coluna tendo seus níveis de agrupamento diferenciados pela quantidade de 'tabs'.
}
\examples{

dataset <- tibble::tribble(
  ~'cidade',~'bairro',~'rua',~'qtd_casas',~'obs',
  'Rio','Gloria','Rua A',41,'',
  'Rio','Catete','Rua B',12,'Historica',
  'Rio','Gloria','Rua C',75,'',
  'São','Socorro','Rua D',9,'Reparos',
  'São','Socorro','Rua E',15,'',
  'São','Saude','Rua F',73,'',
)

beautiful_table(
  dados_brutos = dataset,
  campos_agrupar = c('cidade', 'bairro', 'rua'),
  campos_somar = c('qtd_casas'),
  campos_informacao = c('obs')
)

}
