% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtros.R
\name{cd_municipio_selecionado}
\alias{cd_municipio_selecionado}
\title{Obter Código do Município Selecionado}
\usage{
cd_municipio_selecionado(input, connection, ufs)
}
\arguments{
\item{input}{Um objeto contendo as seleções feitas pelo usuário, como o código do município selecionado.}

\item{connection}{Uma conexão ativa com o banco de dados.}

\item{ufs}{Um vetor de caracteres contendo as siglas das UFs selecionadas.}
}
\value{
Um vetor de caracteres contendo os códigos dos municípios selecionados.
}
\description{
Esta função obtém o código do município selecionado com base nos inputs do usuário, a conexão com o banco de dados e as UFs selecionadas. 
Ela utiliza a tabela "dim_municipio" no banco de dados para reunir os filtros e extrair os códigos de município.
}
