% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtros.R
\name{cd_subclasse_selecionado}
\alias{cd_subclasse_selecionado}
\title{Obter Código da Subclasse Selecionada}
\usage{
cd_subclasse_selecionado(input, connection)
}
\arguments{
\item{input}{Um objeto contendo as seleções feitas pelo usuário, como setores, subsetores, subsubsetores e subclasses selecionadas.}

\item{connection}{Uma conexão ativa com o banco de dados.}
}
\value{
Um vetor de caracteres contendo os códigos das subclasses selecionadas.
}
\description{
Esta função obtém os códigos das subclasses selecionadas com base nos inputs do usuário e na conexão com o banco de dados. 
Ela utiliza a tabela "dim_subclasse" no banco de dados para reunir os filtros e extrair os códigos das subclasses.
}
