% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database_server.R
\name{connect_database}
\alias{connect_database}
\title{Conectar-se ao Banco de Dados SQLite}
\usage{
connect_database()
}
\description{
Esta função estabelece uma conexão com o banco de dados SQLite criado anteriormente.
Ela utiliza a função `dbConnect` do pacote DBI para estabelecer a conexão com o banco de dados.
}
