% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_fatos.R
\name{converte_para_toneladas}
\alias{converte_para_toneladas}
\title{Converte unidades}
\usage{
converte_para_toneladas(dados, origem, dim_conversao)
}
\arguments{
\item{dados}{}

\item{origem}{}

\item{dim_conversao}{}
}
\description{
Utilizando a dim_conversao e os dados passados faz a conversão.
Todos os dados estarão em toneladas na coluna qtd_produzida_t
Origem é um relação entre o produto dos dados e o produto_referência da dim_conversão
Ex. origem <- c("produto_ibge" = "produto_referencia")
}
\examples{
""
}
