% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sqlite.R
\name{create_sqlite}
\alias{create_sqlite}
\title{Criar Banco de Dados SQLite}
\usage{
create_sqlite()
}
\description{
Esta função cria um banco de dados SQLite e popula as tabelas com os dados fornecidos.
Ela utiliza da função `create_tables` para gerar as tabelas e posteriormente insere os dados armazenados em feather nelas. 
Além disso, a função cria índices nas tabelas para otimizar a consulta posterior.
}
