% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sqlite.R
\name{create_tables}
\alias{create_tables}
\title{Criar Tabelas}
\usage{
create_tables(connection)
}
\arguments{
\item{connection}{Um objeto de conexão de banco de dados.}
}
\description{
Esta função cria as tabelas necessárias para um banco de dados específico, nesse caso o banco uitilizados pelo SIEnergia.
Ela contém um comando SQL para criar cada tabela, incluindo suas colunas e restrições. Depois utiliza a função `map` para criar todas as tabelas com a função `dbExecute`
}
\examples{
conn <- DBI::dbConnect(...)
create_tables(conn)

}
