% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_dimensoes.R
\name{cria_dim_combustivel}
\alias{cria_dim_combustivel}
\title{Cria a dimensão combustível}
\usage{
cria_dim_combustivel()
}
\value{
dim_combustivel A dataframe
}
\description{
Trata os arquivos xlsx que contém os dados que compõem a dimensão combustível.
Tem 6 subdivisões:
Anexo 6, 7, 8, 9
Matriz aberta, consolidada
}
\examples{
""
}
