% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_dimensoes.R
\name{cria_dim_data}
\alias{cria_dim_data}
\title{Cria a dimensão data}
\usage{
cria_dim_data()
}
\value{
dim_data A dataframe
}
\description{
Usa a data do sistema e a data fixa de 1974 para criar a dimensão.
Criando informações sobre safra, horizonte(origem informação) e um cd_data para o banco
}
\examples{
""
}
