% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_dimensoes.R
\name{cria_dim_energetico}
\alias{cria_dim_energetico}
\title{Cria a dimensão energético}
\usage{
cria_dim_energetico()
}
\value{
dim_energetico A dataframe
}
\description{
Essa dimensão é feita principalmente manualmente, por isso apenas importamos
uma tabela do excel que contém as informações.
}
