% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_dimensoes.R
\name{cria_dim_municipio}
\alias{cria_dim_municipio}
\title{Cria dimensão município}
\usage{
cria_dim_municipio()
}
\value{
dim_municipio A dataframe
}
\description{
Importa as tabelas de área, divisão territorial do IBGE e um excel com as distribuidoras.
Junta as mesmas pelo código do município e temos as informações em uma só tabela.
}
\examples{
""
}
