% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_dimensoes.R
\name{cria_dim_origem}
\alias{cria_dim_origem}
\title{Cria a dimensão origem}
\usage{
cria_dim_origem()
}
\value{
dim_origem A dataframe
}
\description{
Serve para rastrear a origem do dado em todos os niveis de granularidade.
É feita manualmente com a função tribble
}
\examples{
""
.
}
