% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_dimensoes.R
\name{cria_dim_produto}
\alias{cria_dim_produto}
\title{Cria a dimensão produto}
\usage{
cria_dim_produto()
}
\value{
dim_produto A dataframe
}
\description{
Trata os arquivos csv que contém os dados que compõem a dimensão produto.
Tem 7 subdivisões: 
PAM, 
PPM, 
PEVS, 
ABATE, 
SNIS, 
MAPA
}
\examples{
""
}
