% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_dimensoes.R
\name{cria_dim_subclasse}
\alias{cria_dim_subclasse}
\title{Cria a dimensão subclasse}
\usage{
cria_dim_subclasse()
}
\value{
dim_subclasse A dataframe
}
\description{
Usa os dados do CNAE para gerar secao, divisao, grupo, classe e subclasse
}
\examples{
""
}
