% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_dimensoes.R
\name{cria_dim_uf}
\alias{cria_dim_uf}
\title{Cria dim_uf}
\usage{
cria_dim_uf()
}
\value{
dim_uf A dataframe
}
\description{
Importa um csv contendo diversas informações sobre as UFs brasileiras.
}
\examples{
""
}
