% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_fatos.R
\name{cria_fato_combustivel}
\alias{cria_fato_combustivel}
\title{Cria a fato combustivel}
\usage{
cria_fato_combustivel()
}
\value{
fato_energetico A dataframe
}
\description{
A partir da fato_energetico gera a fato_combustivel
Usa da fator_energetico_combustivel para fazer as conversões e posteriormente
mescla com a dim_combustivel para informações adicionais.
}
\examples{
""
}
