% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_fatos.R
\name{cria_fato_energetico}
\alias{cria_fato_energetico}
\title{Cria a fato energetico}
\usage{
cria_fato_energetico()
}
\value{
fato_energetico A dataframe
}
\description{
A partir da fato_produto_mu gera a fato_energetico.
Usa da fator_produto_energetico para fazer as conversões e posteriormente
mescla com a dim_energetico para informações adicionais.
}
\examples{
""
}
