% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_fatos.R
\name{cria_fato_produto_mu}
\alias{cria_fato_produto_mu}
\title{Cria fato produto municipio}
\usage{
cria_fato_produto_mu()
}
\value{
fato_produto A dataframe
}
\description{
Agrega os dados do IBGE na granularidade de municipio,
esses são os agropecuários do SIDRA e resíduos do SNIS.
}
\examples{
""
}
