% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_fatos.R
\name{cria_fato_produto_uf}
\alias{cria_fato_produto_uf}
\title{Cria fato produto UF}
\usage{
cria_fato_produto_uf()
}
\value{
fato_produto A dataframe
}
\description{
Agrega os dados do IBGE na granularidade de UF, apenas abate,
com os dados da fato_produto_mu já somados por UF para igualar a granularidade.
}
\examples{
""
}
