% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_fatores.R
\name{cria_fator_cap3}
\alias{cria_fator_cap3}
\title{Cria o fator cap3}
\usage{
cria_fator_cap3()
}
\value{
dim_cap3 A dataframe
}
\description{
Utiliza a tabela do BEN para gerar a dimensão cap3 contendo os setores
}
\examples{
""
}
