% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_fatores.R
\name{cria_fator_energetico_combustivel}
\alias{cria_fator_energetico_combustivel}
\title{Cria fator de conversão entre energético e combustível}
\usage{
cria_fator_energetico_combustivel()
}
\value{
fator_energetico_combustivel A dataframe
}
\description{
Dado que a criação dos fatores e quais energeticos serão convertidos é manual,
essa função apenas busca o excel que contém esses dados.
As informações são mantidas no excel pois devem estar em constante mudança.
}
\examples{
""
}
