% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_fatores.R
\name{cria_fator_produto}
\alias{cria_fator_produto}
\title{Cria fator de conversão dos produtos}
\usage{
cria_fator_produto()
}
\value{
fator_produto A dataframe
}
\description{
Dados provenientes de um PDF do ibge e passado a mão.
Tendo a tabela escrita esta é expandida para os anos de estudo.
}
\examples{
""
}
