% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_fatores.R
\name{cria_fator_produto_energetico}
\alias{cria_fator_produto_energetico}
\title{Cria fator de conversão entre produto e energético}
\usage{
cria_fator_produto_energetico()
}
\value{
fator_produto_energetico A dataframe
}
\description{
Dado que a criação dos fatores e quais produtos serão convertidos é manual,
essa função apenas busca o excel que contém esses dados.
As informações são mantidas no excel pois devem estar em constante mudança.
}
\examples{
""
}
