% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suporte_ui.R
\name{cria_filtros_sem_rota}
\alias{cria_filtros_sem_rota}
\title{Cria Filtros sem Rota}
\usage{
cria_filtros_sem_rota(id)
}
\arguments{
\item{id}{O ID usado para nomear os elementos de entrada evitando conflito entre modulos.}
}
\value{
Um objeto `html` a ser usado na UI do aplicativo shiny.
}
\description{
Esta função cria uma faixa de filtros, sem conter filtro de rota, utilizada no aplicativo Shiny.
Dentro do SIEnergia Viewer é usada nos paineis de produto e energético visto que esses não são afetados pelo filtro de rotas.
Note que essa função apenas faz o esqueleto dos filtros e não os preenchem com valores.
}
