% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl.R
\name{cria_tabela_metadados}
\alias{cria_tabela_metadados}
\title{Cria a tabela de metadados}
\usage{
cria_tabela_metadados()
}
\value{
tabela_metadados
}
\description{
Utilizada principalmente para as labels das colunas
}
\examples{
""
}
