% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabelas.R
\name{cria_tabela_reactable}
\alias{cria_tabela_reactable}
\title{Cria uma tabela Reactable}
\usage{
cria_tabela_reactable(
  dados_tabela,
  columns_list = NULL,
  groups_list = NULL,
  background_color = "rgb(12, 35, 64)",
  pagination = FALSE,
  sortable = FALSE
)
}
\arguments{
\item{dados_tabela}{Um conjunto de dados contendo os dados da tabela.}

\item{columns_list}{Uma lista de colunas personalizadas para a tabela (opcional).}

\item{groups_list}{Uma lista de grupos de colunas para a tabela (opcional).}

\item{background_color}{A cor de fundo dos cabeçalhos da tabela (padrão: "rgb(12, 35, 64)").}

\item{pagination}{Um valor lógico indicando se a paginação deve ser ativada (padrão: FALSE).}

\item{sortable}{Um valor lógico indicando se a ordenação das colunas deve ser ativada (padrão: FALSE).}
}
\value{
Uma tabela Reactable com as configurações fornecidas.
}
\description{
Essa função cria uma tabela Reactable com base nos dados fornecidos.
Ela aceita um conjunto de dados dados_tabela como entrada e também oferece a opção de personalizar as colunas e grupos de colunas da tabela, fornecendo as listas columns_list e groups_list, respectivamente.
Além disso, é possível definir a cor de fundo dos cabeçalhos da tabela usando o parâmetro background_color (padrão: "rgb(12, 35, 64)").
A função também permite ativar a paginação e a ordenação das colunas por meio dos parâmetros pagination e sortable.
}
