% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suporte_ui.R
\name{cria_well_panel_modelo}
\alias{cria_well_panel_modelo}
\title{Cria Well Panel Modelo}
\usage{
cria_well_panel_modelo(id)
}
\arguments{
\item{id}{O ID usado para nomear os elementos de saída.}
}
\value{
Um objeto `html` a ser usado na UI do aplicativo shiny.
}
\description{
Esta função cria o corpo principal do painel modelo do aplicativo Shiny, contendo os gráficos, tabelas e botões de download.
}
