% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_fatos.R
\name{dados_censo_pecuaria}
\alias{dados_censo_pecuaria}
\title{Censo pecuária}
\usage{
dados_censo_pecuaria()
}
\description{
Tem como função criar um dataframe que relacione os municipios com uma lotação.
Para isso calcula a lotação por municipio com os dados do censo.
}
\examples{
""
}
