% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_fatos.R
\name{dados_sidra_abate_uf}
\alias{dados_sidra_abate_uf}
\title{Importa dados SIDRA}
\usage{
dados_sidra_abate_uf()
}
\value{
fato_sidra_uf A dataframe
}
\description{
Criado principalmente para tratar os dados de ABATE que tem granularidade apenas em UF.
Importa diversos csv contendo as informações do sidra por unidade federal.
Compila os dados em um único dataframe e o mescla com a dim_produto, dim_municipio e dim_subclasse,
a fim incluir os códigos deles.
}
\examples{
""
}
