% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_fatos.R
\name{dados_sidra_agro_municipio}
\alias{dados_sidra_agro_municipio}
\title{Importa dados SIDRA}
\usage{
dados_sidra_agro_municipio()
}
\value{
fato_sidra_mu A dataframe
}
\description{
Importa diversos csv contendo as informações do sidra por municipio.
Compila os dados em um único dataframe e o mescla com a dim_produto, dim_municipio e dim_subclasse,
a fim incluir os códigos deles.
Contém os dados da PAM, PPM e PEVS
}
\examples{
""
}
