% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_fatos.R
\name{dados_snis_municipio}
\alias{dados_snis_municipio}
\title{Importa dados SNIS}
\usage{
dados_snis_municipio()
}
\value{
fato_snis A dataframe
}
\description{
Trata um csv contendo as informações de produção de resíduos.
}
\examples{
""
}
