% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_dimensoes.R
\name{depara_etl_geral}
\alias{depara_etl_geral}
\title{Depara de dados}
\usage{
depara_etl_geral(dado_origem, dado_destino, depara, sheet)
}
\arguments{
\item{dado_origem}{Uma coluna do dataframe que será tratada}

\item{dado_destino}{Nome da coluna2 na planilha}

\item{depara}{Um vetor nomeado para ligar o dado bruto à planilha}

\item{sheet}{Qual sheet contém o tratamento}
}
\value{
resultado Uma coluna de dataframe
}
\description{
Serve para o tratamento específico de alguns dados.
Faz um depara através de uma planilha excel (coluna1 = dado_bruto, coluna_2 = dado_tratado).
A planilha excel contém diversas sheets a fim de podermos usarmos essa função de forma geral.
}
\examples{
""
}
