% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{gera_grafico_barra}
\alias{gera_grafico_barra}
\title{Geração de gráfico de barras interativo}
\usage{
gera_grafico_barra(dados, campo_contagem, y_label = NULL, tooltip)
}
\arguments{
\item{dados}{Um data frame contendo os dados a serem utilizados no gráfico.}

\item{campo_contagem}{O nome da coluna que contém a variável a ser quantificada para cada categoria.}

\item{y_label}{Um rótulo opcional para o eixo Y do gráfico. Se não for fornecido, será usado o rótulo padrão.}

\item{tooltip}{Uma string que define o conteúdo do tooltip exibido quando o usuário passa o mouse sobre as barras.}
}
\value{
Um gráfico de barras interativo.
}
\description{
Esta função gera um gráfico de barras interativo com base nos dados fornecidos,
onde cada barra representa uma categoria e a altura da barra representa a contagem
de uma variável específica.
}
