% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{gera_grafico_linha}
\alias{gera_grafico_linha}
\title{Geração de gráfico de linha interativo}
\usage{
gera_grafico_linha(dados, campo_contagem, y_label)
}
\arguments{
\item{dados}{Um data frame contendo os dados a serem utilizados no gráfico.}

\item{campo_contagem}{O nome da coluna que contém os valores a serem plotados ao longo dos anos.}

\item{y_label}{Um rótulo opcional para o eixo Y do gráfico. Se não for fornecido, será usado o rótulo padrão.}
}
\value{
Um gráfico de linha interativo.
}
\description{
Esta função gera um gráfico de linha interativo com base nos dados fornecidos,
onde cada linha representa uma categoria e os pontos representam os valores dessa categoria ao longo dos anos.
}
