% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{gera_grafico_mosaico}
\alias{gera_grafico_mosaico}
\title{Gerar Gráfico de Mosaico}
\usage{
gera_grafico_mosaico(
  dados,
  area,
  categorizacao,
  tooltip = NULL,
  height = NULL,
  porcentagem = FALSE
)
}
\arguments{
\item{dados}{Um dataframe contendo os dados a serem plotados no gráfico de mosaico.}

\item{area}{O nome da coluna que representa a área dos retângulos no gráfico de mosaico.}

\item{categorizacao}{O nome da coluna que será usada para categorizar os retângulos no gráfico de mosaico.}

\item{tooltip}{Uma string contendo o texto que será exibido como tooltip para cada retângulo. Pode conter placeholders como "{area}" para representar o valor da área.}

\item{height}{A altura do gráfico de mosaico. Se não for especificada, a altura padrão será usada.}

\item{porcentagem}{Um valor lógico indicando se a área deve ser exibida como porcentagem. Se for TRUE, a área será multiplicada por 100.}
}
\value{
Um objeto girafe contendo o gráfico de mosaico.
}
\description{
Esta função gera um gráfico de mosaico a partir dos dados fornecidos. 
O gráfico de mosaico exibe a distribuição de uma variável em áreas retangulares proporcionais ao valor dessa variável. 
O tamanho dos retângulos é determinado pela área especificada nos dados.
}
