% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{gera_mapa_micros}
\alias{gera_mapa_micros}
\title{Gerar Mapa das Microrregiões}
\usage{
gera_mapa_micros(dados, code_muni_local = NULL, fill, titulo = NULL)
}
\arguments{
\item{dados}{Um objeto sf contendo os dados geográficos das microrregiões.}

\item{code_muni_local}{Um vetor com os códigos dos municípios que determinam as microrregiões a serem exibidas no mapa. Se não for fornecido, todas as microrregiões serão exibidas.}

\item{fill}{O nome da coluna que será usada para preencher as microrregiões com cores diferentes.}

\item{titulo}{O título do mapa.}
}
\value{
Um objeto girafe contendo o mapa das microrregiões.
}
\description{
Esta função gera um mapa das microrregiões a partir dos dados fornecidos. 
O mapa exibe as microrregiões coloridas de acordo com a variável especificada em `fill`.
}
