% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{gera_mapa_municipios}
\alias{gera_mapa_municipios}
\title{Gerar Mapa dos Municipios}
\usage{
gera_mapa_municipios(
  dados,
  code_muni_local = NULL,
  fill,
  titulo = NULL,
  tooltip = NULL
)
}
\arguments{
\item{dados}{Um objeto sf contendo os dados geográficos dos mmunicipios.}

\item{code_muni_local}{Um vetor com os códigos dos municípios que determinam os municipios a serem exibidas no mapa. Se não for fornecido, todas os municipios serão exibidos.}

\item{fill}{O nome da coluna que será usada para preencher as municipios com cores diferentes.}

\item{titulo}{O título do mapa.}

\item{tooltip}{Uma string contendo o texto que será exibido como tooltip para cada municipio.}
}
\value{
Um objeto girafe contendo o mapa dos municipios
}
\description{
Esta função gera um mapa dos municipios a partir dos dados fornecidos. 
O mapa exibe os municipios coloridas de acordo com a variável especificada em `fill`.
}
