% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{gera_mapa_ufs}
\alias{gera_mapa_ufs}
\title{Gerar Mapa das Unidades Federativas (UFs)}
\usage{
gera_mapa_ufs(
  dados,
  ufs_selecionadas = NULL,
  fill,
  tooltip,
  titulo,
  bbox = NULL
)
}
\arguments{
\item{dados}{Um objeto sf contendo os dados geográficos das UFs.}

\item{ufs_selecionadas}{Um vetor contendo as siglas das UFs a serem exibidas no mapa. Se não for especificado, todas as UFs presentes nos dados serão exibidas.}

\item{fill}{O nome da coluna que será usada para colorir as UFs no mapa.}

\item{tooltip}{Uma string contendo o texto que será exibido como tooltip para cada UF.}

\item{titulo}{O título do mapa.}

\item{bbox}{Uma lista com as coordenadas do bounding box do mapa. Se não for especificado, o bounding box será calculado automaticamente com base nas UFs selecionadas.}
}
\value{
Um objeto girafe contendo o mapa interativo das UFs.
}
\description{
Esta função gera um mapa interativo das Unidades Federativas (UFs) a partir dos dados fornecidos. 
O mapa exibe as microrregiões coloridas de acordo com a variável especificada em `fill`.
}
