% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{padronizacao_tabela_evolucao}
\alias{padronizacao_tabela_evolucao}
\title{Função auxiliar à tabela evolução}
\usage{
padronizacao_tabela_evolucao(ano = 2021)
}
\arguments{
\item{ano}{ano a ser feita a instanciação, deve ser o ultimo ano a ser mostrado na tabela}
}
\value{
tibble com instanciação de todos os 'n3_grupo'
}
\description{
Na tabela evolução de cada painel temos uma coluna para cada cultura.
Devido a natureza do Reactable não podemos gerar as colunas reativamente e por isso temos essa função.
A função apenas gera uma instancia de cada 'n3_grupo' para que a Reactable possa gerar a tabela sem dar o erro de falta de dados.
}
