% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{poligonos_estados_simples}
\alias{poligonos_estados_simples}
\title{Poligonos de estados}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 27 rows and 6 columns.
}
\usage{
poligonos_estados_simples
}
\description{
Poligonos de estados
}
\keyword{datasets}
