% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{poligonos_microrregiao_simples}
\alias{poligonos_microrregiao_simples}
\title{Poligonos de microrregioes}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 556 rows and 3 columns.
}
\usage{
poligonos_microrregiao_simples
}
\description{
Poligonos de microrregioes
}
\keyword{datasets}
