% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{poligonos_municipios_simples}
\alias{poligonos_municipios_simples}
\title{Poligonos de municipios}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 5566 rows and 5 columns.
}
\usage{
poligonos_municipios_simples
}
\description{
Poligonos de municipios
}
\keyword{datasets}
