% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{populacao_tabela_municipio}
\alias{populacao_tabela_municipio}
\title{População por município}
\usage{
populacao_tabela_municipio(
  dados,
  connection,
  inicio_periodo,
  fim_periodo,
  campos_informacao
)
}
\arguments{
\item{dados}{Um data frame contendo informações sobre os municípios.}

\item{connection}{Uma conexão com o banco de dados que contém a tabela da população.}

\item{inicio_periodo}{A data de início do período desejado para filtrar os dados da população.}

\item{fim_periodo}{A data de término do período desejado para filtrar os dados da população.}

\item{campos_informacao}{Um vetor com os nomes dos campos de informação adicionais a serem incluídos na saída.}
}
\value{
Um data frame com informações sobre a população por município, incluindo os campos de informação especificados.
}
\description{
Esta função retorna uma tabela com informações sobre a população de cada município,
incluindo dados demográficos e campos adicionais fornecidos.
}
