% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl.R
\name{run_etl_atualizado}
\alias{run_etl_atualizado}
\title{Roda o etl}
\usage{
run_etl_atualizado()
}
\value{
void O objetivo da função é escrever os feather, não retornar algo
}
\description{
Chama as funções que tratam os dados brutos e retornam dataframes com os mesmos,
e então salva esses dados no formato .feather
}
\examples{
""
}
