% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sigla_abrev}
\alias{sigla_abrev}
\title{Sigla Abreviada}
\usage{
sigla_abrev(x, accuracy = 3)
}
\arguments{
\item{x}{O número a ser convertido em sigla abreviada.}

\item{accuracy}{O número de dígitos decimais a serem arredondados (padrão: 3).}
}
\value{
Uma string contendo o numero com a sigla de magnitude.
}
\description{
Esta função reescreve um número adicionando uma sigla abreviada correspondente, indicando a ordem de magnitude.
}
\examples{
"sigla_abrev(10000) == '10 mil'"
}
