% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtros.R
\name{sigla_ufs_selecionadas}
\alias{sigla_ufs_selecionadas}
\title{Obter Siglas das UF Selecionadas}
\usage{
sigla_ufs_selecionadas(input, connection)
}
\arguments{
\item{input}{Um objeto contendo as seleções feitas pelo usuário, como as regiões selecionadas e as UFs selecionadas.}

\item{connection}{Uma conexão ativa com o banco de dados.}
}
\value{
Um vetor de caracteres contendo as siglas das UFs selecionadas.
}
\description{
Esta função obtém as siglas dos estados (UFs) que foram selecionados com base nos inputs de regiões e UFs feitos pelo usuário (especificamente input$mosaico_regioes_selected e input$mapa_brasil). 
Ela utiliza a tabela "dim_uf" no banco de dados para reunir os filtros e extrair as siglas.
}
