% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tema_header_footer_reactable}
\alias{tema_header_footer_reactable}
\title{Cria um tema personalizado para cabeçalhos e rodapés de tabelas Reactable}
\usage{
tema_header_footer_reactable(color)
}
\arguments{
\item{color}{A cor de fundo dos cabeçalhos e rodapés da tabela.}
}
\value{
Um objeto de tema personalizado para tabelas Reactable.
}
\description{
Esta função cria um tema personalizado para os cabeçalhos e rodapés de tabelas Reactable, definindo o estilo dos cabeçalhos e rodapés.
}
