% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{texto_tooltip}
\alias{texto_tooltip}
\title{Texto para Tooltip}
\usage{
texto_tooltip(dados_grafico, categoria, info)
}
\arguments{
\item{dados_grafico}{Um conjunto de dados contendo as informações para o gráfico.}

\item{categoria}{A coluna que representa a categoria para agrupar os dados.}

\item{info}{A coluna que contém as informações a serem exibidas no tooltip.}
}
\value{
Um conjunto de dados com uma coluna adicional contendo o texto formatado para o tooltip.
}
\description{
Esta função cria o texto para exibição no tooltip de um gráfico, com base nos dados, na categoria e nas informações fornecidas.
}
